+++ tests/test_introspection_auth.py
@@ -96,7 +96,7 @@ class TestTokenIntrospectionAuth(TestCase):
             user=self.resource_server_user, token="12345678900",
             application=self.application,
             expires=timezone.now() + datetime.timedelta(days=1),
-            scope="read write introspection"
+            scope="introspection"
         )
 
         self.invalid_token = AccessToken.objects.create(
@@ -129,7 +129,8 @@ class TestTokenIntrospectionAuth(TestCase):
         token = self.validator._get_token_from_authentication_server(
             self.resource_server_token.token,
             oauth2_settings.RESOURCE_SERVER_INTROSPECTION_URL,
-            oauth2_settings.RESOURCE_SERVER_AUTH_TOKEN
+            oauth2_settings.RESOURCE_SERVER_AUTH_TOKEN,
+            oauth2_settings.RESOURCE_SERVER_INTROSPECTION_CREDENTIALS
         )
         self.assertIsNone(token)
 
@@ -141,7 +142,8 @@ class TestTokenIntrospectionAuth(TestCase):
         token = self.validator._get_token_from_authentication_server(
             "foo",
             oauth2_settings.RESOURCE_SERVER_INTROSPECTION_URL,
-            oauth2_settings.RESOURCE_SERVER_AUTH_TOKEN
+            oauth2_settings.RESOURCE_SERVER_AUTH_TOKEN,
+            oauth2_settings.RESOURCE_SERVER_INTROSPECTION_CREDENTIALS
         )
         self.assertIsInstance(token, AccessToken)
         self.assertEqual(token.user.username, "foo_user")